More kwarg handling

+++ quark/drivers/security_groups.py
@@ -28,10 +28,10 @@ class SecurityGroupDriver(object):
         if "security_groups" in kwargs:
             device_id = kwargs.get('device_id')
             mac_address = kwargs.get('mac_address')
+            if not device_id or not mac_address:
+                LOG.warning('device_id or mac_address not given, ignored.')
+                return
             if kwargs["security_groups"]:
-                if not device_id or not mac_address:
-                    LOG.warning('device_id or mac_address not given, ignored.')
-                    return
                 payload = client.serialize_groups(
                     kwargs["security_groups"])
                 client.apply_rules(device_id, mac_address, payload)
@@ -44,6 +44,9 @@ class SecurityGroupDriver(object):
         try:
             device_id = kwargs.get('device_id')
             mac_address = kwargs.get('mac_address')
+            if not device_id or not mac_address:
+                LOG.warning('device_id or mac_address not given, ignored.')
+                return
             client.delete_vif(device_id, mac_address)
         except Exception:
             LOG.exception("Failed to reach the security groups backend")
